eesa-backend/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
frozenlist==1.3.3
fuzzywuzzy==0.18.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
kaitaistruct==0.10
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.4
outcome==1.2.0
pandas==2.0.0
passlib==1.7.4
playsound==1.3.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pygame==2.4.0
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
selenium==4.9.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
twilio==8.1.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
wsproto==1.2.0
yarl==1.9.1
zipp==3.15.0
zstandard==0.21.0
